requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.1
django-widget-tweaks==1.5.0
psycopg2-binary==2.9.9
rules==3.5
sqlparse==0.5.1
typing_extensions==4.12.2
djangorestframework==3.14.0
